Move module to `package mill.contrib.webpack`
