+++ contrib/webpack/test/resources/customWpConfig.js
