+++ contrib/webpack/test/resources/package-simple.json
