+++ contrib/webpack/test/resources/package-transitive.json
