+++ contrib/webpack/test/src/mill/contrib/webpack/ScalaJSWebpackTests.scala
@@ -1,12 +1,11 @@
-package mill
-package contrib
+package mill.contrib.webpack
 
 import ammonite.ops
-import ammonite.ops.Path
-import mill.contrib.ScalaJSWebpackModule._
 import mill.define.Target
 import mill.scalalib._
 import mill.util.{TestEvaluator, TestUtil}
+import mill.{Agg, T}
+import os.ReadablePath
 import utest.framework.TestPath
 import utest.{TestSuite, Tests, _}
 
@@ -38,14 +37,12 @@ object ScalaJSWebpackTests extends TestSuite {
     f(ev)
   }
 
-  val resPath: Path =
-    ops.pwd / "contrib" / "webpacklib" / "test" / "resources"
   val sProps: String =
-    ops.read(resPath / "snabbdom-custom-props.js")
+    ops.read(os.resource / "snabbdom-custom-props.js")
   val npmSimpleJson: String =
-    ops.read(resPath / "package-simple.json")
+    ops.read(os.resource / "package-simple.json")
   val npmTransitiveJson: String =
-    ops.read(resPath / "package-transitive.json")
+    ops.read(os.resource / "package-transitive.json")
   val wpConf: String => String = p =>
     s"""// Webpack config generated by ScalaJSWebpackModule
        |const generatedWebpackCfg = {
@@ -154,7 +151,7 @@ object ScalaJSWebpackTests extends TestSuite {
         result(
           ev.outPath,
           "webpack.config.js",
-          Some(resPath / "customWpConfig.js"),
+          Some(os.resource / "customWpConfig.js": ReadablePath),
           "out.js",
           "out-bundle.js",
           false
@@ -165,7 +162,7 @@ object ScalaJSWebpackTests extends TestSuite {
         assert(
           cfg == mergedWithCustomWpConf(
             ev.outPath.toString,
-            (resPath / "customWpConfig.js").toString),
+            (os.resource / "customWpConfig.js").toString),
         )
       }
